review input
